requirements.txt
langchain
langchain_community
langchain-openai
openai
streamlit
python-dotenv
wikipedia
numexpr
icecream